Share the LDS capacity lookup with any FlyDSL kernel
How much LDS a workgroup gets is a property of the device, not of the grouped
GEMM, and any kernel validating its own budget before tracing needs the same
number. Move it next to the other shared FlyDSL helpers; what stays behind is
the part that is specific, which is how this kernel's tiles consume it.
